Replace getPageSchema
