+++ front/app/components/Form/Components/Layouts/utils.ts
@@ -9,12 +9,7 @@ import {
 import { forOwn, isEmpty } from 'lodash-es';
 
 import { FormData } from '../../typings';
-import {
-  ExtendedRule,
-  ExtendedUISchema,
-  extractElementsByOtherOptionLogic,
-  isVisible,
-} from '../Controls/visibilityUtils';
+import { ExtendedRule, ExtendedUISchema } from '../Controls/visibilityUtils';
 
 export interface PageType extends Layout {
   type: 'Page';
@@ -82,42 +77,6 @@ export const getSanitizedFormData = (data) => {
   });
 
   return sanitizedFormData;
-};
-
-export const getPageSchema = (
-  schema: JsonSchema,
-  pageCategorization: PageType,
-  data: any
-) => {
-  const currentPageElementNames = pageCategorization.elements.map(
-    (uiSchemaElement) => uiSchemaElement.scope.split('/').pop()
-  );
-
-  const pageSchemaProperties = {};
-  Object.entries(schema.properties || {}).forEach(([key, value]) => {
-    if (currentPageElementNames.includes(key)) {
-      pageSchemaProperties[key] = value;
-    }
-  });
-
-  const hiddenElements = pageCategorization.elements
-    .filter((pageElement) => pageElement.ruleArray)
-    .filter((pageElementWithRule) => {
-      return !isVisible(pageElementWithRule, data);
-    })
-    .concat(extractElementsByOtherOptionLogic(pageCategorization, data, true))
-    .map((element) => element['scope'].split('/').pop());
-
-  const pageSchema = Object.assign({}, schema, {
-    required: (schema.required || [])
-      .filter((requiredElementName) =>
-        currentPageElementNames.includes(requiredElementName)
-      )
-      .filter((requiredElement) => !hiddenElements.includes(requiredElement)),
-    properties: pageSchemaProperties,
-  });
-
-  return pageSchema;
 };
 
 export const isPageCategorization: Tester = and(
